Re-factor of the way how arguments are handled.
This is a new approach to handling arguments passed to the newfunction
either from within the Puppet manifest or a template.  This will allow
for passing arguments in a comma-separated fashion in both of the cases
with a guarantee that it will always do the right thing to when parsing.

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>