Resolve issues with compatibility between Rubies 1.8.x and 1.9.x.
This is to make sure that this function will work in older and
more recent versions of Ruby.  We monkey-patch class String when
needed to do so e.g. Ruby 1.8.6 and older to ensure, that the
method String#bytes is always present.

Hint: Ruby 1.8.x and 1.9.x have different way of handling and
dealing with strings / character data: byte-stream versus proper
objects that are encoding-aware ...

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>