Re-factor range function, add zero-padding, step, etc.
This is a complete re-factor of the range function that adds
zero-padding handling, better handling of single integer values,
support for negative integer values, support for strings and
ability to step through values (meaning: every n-th value).

Also, the invocation is different in this version, as now
function requires two arguments (with third being optional).

Signed-off-by: Krzysztof Wilczynski <krzysztof.wilczynski@linux.com>